requirements.txt
wheel-0.34.2
pbkdf2-1.3
aiohttp-3.6.2
aiodns-2.0.0
getmac-0.8.2